refactor landing page components
