+++ src/components/ReadyToTalk/readyToTalk.module.css
@@ -1,8 +1,3 @@
-/* <!-- Start Ready To Talk Area --> */
-
-/*================================================
-Ready To Talk CSS
-=================================================*/
 .ready_to_talk {
 	text-align: center;
 	position: relative;
@@ -140,4 +135,9 @@ Ready To Talk CSS
 	line-height: 1.5;
 	border-radius: 0.25rem;
 }
-/* Ready To Talk Section Ends Here */
+
+.centered {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+}
